Switched PortUtil to use getLoopbackAddress
We reuse the MemcachedDaemonFactory within our application tests.
However, when run from our corporate network the PortUtil fails to
bind to any addresses because of the VPN being used.

Under the hood the VPN doesn't resole InetAddress.getLocalHost() to
the address of your local machine and hence fails to bind ports.

I've updated it to using InetAddress.getLoopbackAddress() to overcome
this.